Migrated to Python3. Works with pypy3.
/web/pings is not working yet, the list comprehension that is used is
buggy in 3.x (https://bugs.python.org/issue10544) so it needs to be
changed to for loops, but I want to get this code out for testing.

Removed the UPnP code as it is extreamly old and there is no good SOAP
clients that don't have a lot of dependencies. Removing for now, may put
it back in later. For now, just manually forward your ports from your
router (most people are probably already doing this).

Works with peers that are using the Python2 version. Has the stripped
out shares for share version 34.